+++ include/scriptos/iterator.hpp
@@ -6,7 +6,7 @@ template <typename T>
 class Iterator
 {
 public:
-    Iterator(T *ptr, u64 stride)
+    Iterator(T *ptr, u32 stride)
         : m_Ptr(ptr), m_Stride(stride) {}
 
     T &operator*() const { return *m_Ptr; }
@@ -29,5 +29,5 @@ public:
 
 private:
     T *m_Ptr;
-    u64 m_Stride;
+    u32 m_Stride;
 };
